Listen on *:port when the port is specified
The expectation should be that if you're listening to a non-default port
you are forwarding the connection in some way and listening on the
IP from DNS won't be useful.

Also print the {ip}:port that's being listened to.

This should help on #39; the issue I ran into was that lego could
listen on public-ip:port (when I specified the port), but I was
forwarding the connection to private-ip:port.