Refactor DNS check
* Implements a (semi-)recursive query mechanism to find all authoritative
nameservers for the zone enclosing the domain.
This is necessary because Google Public DNS
does sometimes not return any NS at all or those of a different zone (e.g.
when the domain is a CNAME pointing elsewhere).
* All authoritative nameservers are then queried for the TXT record to make
sure that it has been completely propagated.
* The waitFor utility function now takes an interval parameter and returns
the last seen error on timeout.