CLI: Reorganize flags
This commit renames the --http flag to --http-address and --tls to
--tls-address.

The --webroot flag is removed and instead webroot is specified by
using --http webroot. The webroot path is now given in an environment
variable, in the same style as the dns credentials.

An httphelp command is added to mirror the dnshelp command.