acme: Export separate authorization and certificate generation steps
Export the following functions: getChallenges, solveChallenges,
requestCertificateForCsr, and requestCertificate (the last one for
consistency). Also, as a consequence of these functions being exported,
authorizationResource has been exported as well.

This allows for more granular use of the API, namely the de-coupling of
the authorization process from the certificate request process.