Implement acme.GetenvOrFile
GetenvOrFile will first attempt to read a variable from the environment
using os.Getenv. Failing that it will see if a variable with "_FILE"
suffixed on the end exists. If it finds such a variable, it will read
the referenced file and return its contents as the value.

This functionality will permit those using Docker Swarm to mount their
secrets (which currently have to come in as files) and then use
variables like CLOUDFLARE_API_KEY_FILE to tell lego where to find the
secret.