Allow for a delay prior to the DNS precheck
This requires the challenge objects to be able to group authorizations
by the order that spawned them, so that waiting isn't done per
authorization, but per order (and only once per order), as all of an
order's `PreSolve` steps are done before all of its `Solve` steps.  This
leads to an `acme.ExtendedAuthorization` object which has a handle to
the `acme.ExtendedOrder`, but then requires threading throughout the
code.