Use provided ZoneName instead of domain
For finding the correct hosting.de-ZoneConfig the ZoneName provided via
HOSTINGDE_ZONE_NAME needs to be used in a ZoneConfigsFindRequest. When
using a domain provided with --domains instead, the API-request returns
an empty ZoneConfig if the ZoneName is different from the exact domain
requested (eg. requesting a cert for something.yourdomain.com with
zoneName yourdomain.com).