feat: ease operation behind proxy servers
When running behind proxy servers (e.g. Apache or NGINX), the
http01.ProviderServer fails when validating the incoming request's
Host header with the domain it presents the key auth for. This is
because the Host header actually contains an internal value (for
Apache it's the host name of the ProxyPass directive), and the
original Host value is moved to "X-Forwarded-Host".

An easy workaround would be to remove that validation, but it has
a justifiable purpose (DNS rebind attack prevention).

This adds an extension to http01.ProviderServer to setup alternative
header names. Aside from "Host" (the previous and current default),
some arbitrary header names (likely "X-Forwarded-Host") it also adds
support for RFC7239 (i.e. "Forwardable" header). In the latter case,
the semantics are a bit different, and actually require a proper
parser.

Also, this behaviour can be controlled with an --http.proxy-header
CLI flag.