cloudflare: handle restricted API tokens
This introduces a separation between updating DNS records and retrieving
the Zone ID for a domain name (in order to perform the DNS update in the
first place).

The (*dns/cloudflare).DNSProvider gains a new API client for this purpose,
which can be configured with a different API token (CF_ZONE_API_TOKEN).
The default is to reuse the client used previously for this operation.
For consistency, its env config variable is renamed to CF_DNS_API_TOKEN.

This commit also introduces a cache for ZoneIDByName() calls, as briefly
discussed in #984. It reduces the number of calls to Cloudflare's zones
endpoint.

fixes #984