dns/rfc2136: fix provider for legacy DynDNS Pro accounts
The DNS provider for legacy DynDNS accounts does not seem to
respond properly to SOA requests, which breaks the zone apex
detection.

This commit gives the user a chance to skip the SOA lookup
and directly provide the name for the primary DNS server
(also known as SOA MNAME).

Fixes #1186.

Signed-off-by: Dominik Menke <dom@digineo.de>