pdns: Allow setting the server name.
When using the API straight on the auth server, the servername in the
API is 'localhost'. However, when fronted with a multi-server proxy,
the name is usually different[1].

This commit adds the `PDNS_SERVER_NAME` variable to set the servername.
If it is not set, it is 'localhost'. Hence, this change is fully
backward compatible.

1 - https://doc.powerdns.com/authoritative/http-api/server.html