Fix bug with non-leaf domain support in vinylDNS provider.
The VinylDNS backend has an issue with dotted hostname governance when a SOA record matches one of the requested ACME domains.  It will accept the RecordSet changes, but will fail to create them silently in the background.  This solution leverages the native lego functions to obtain the SOA in which the RecordSet should be placed rather than assuming it's always needs to use a dotted hostname.