hurricane: only treat first word of response body as response code
In the hurricane DNS provider, consider only the first word in the
response body to be the response code when checking whether the request
succeeded; specifically, do not necessarily consider multi-word
responses to be indicative of server errors. This corrects the handling
of `nochg` responses from he.net's Dynamic DNS service, in which the
response code is followed by the unchanged value of the TXT record.

Fixes #1408.