provider:dns: index supported provider list via a custom type
A new type is introduced: `SupportedProvider` - which is an alias to int.

And array of const over iota is used to set the value of every provider.

The type implement the interface Stringer + the MarshalJSON / UnmarshalJSON one,
useful for json (de-)serialization, or even cli flag feature (see
 github.com/thediveo/enumflag).

A special error is yield in case of invalid value provided to set
the type: `ErrUnsupportedProvider`.

A factory signature is introduce (`providerFactory`, which is an alias to
`func() (challenge.Provider, error)`) to dump the switch case in favor of an
index over a `map[supportedProvder]ProviderFactory`.

two other new methods: `IsProviderSupported` and `GetSupportedProvider`