Fix EdgeDNS Present and CleanUp logic
The current Present logic of the EdgeDNS provider tries to create the TXT record even if it has just updated the existing TXT record.
The current CleanUp logic of the EdgeDNS provider deletes the TXT record even if it found that there are still challenges in the TXT record that should remain there.

This commit is to fix these issues.