review: get local address from dns.Server
By setting the port to 0, we defer the actual port selection
to the OS (and thus we're guaranteed to use a free port).

This eliminates an external dependency (jsumners/go-getport), which
tried to do the same, but has a race condition between it returning
an unused socket address, and us using that address in our tests.

When instructing dns.Server to use port 0, the source for the local
port depends on the protocol: for UDP this is (*dns.Server).PacketConn,
and for TCP we find it in (*dns.Server).Listener.