review: improve test failure scenarios
In general, require.NoError() should be preferred to assert.NoError(),
especially if further assertions follow after that.

The stretchr/testify/assert package does not call (*testing.T).FailNow()
when the assertion fails (and stop the test), which can lead to pretty
unreadable error message, e.g. when the rest of the test code panics.