review: preserve public API
Introducing a new parameter to an exported (public) function changes
the API and requires a major version bump. Remember that Lego is used
as library, e.g. as part of the Traefic app proxy.

It is possible to preserve the API by extending the interface on
the ProviderServer structs though.