nearlyfreespeech: fix authantication
This change fixes the issue where the path that was hashed was missing
the initial slash. There is a bug in Golang's stdlib that surfaces when
one uses `url.Parse` with a URL that has an empty path and no slash just
after the domainname (https://golang.org/issue/58605). We add a
workaround for that issue.