nearlyfreespeech: test values of the authentication header
This change adds a golden value test for the contents of the
authentication header, to guard against changes caused by e.g. updates
to Golang stdlib. Sadly, we don't have a reference implementation to
test against, so these must be golden-style tests that require the
golden values to be updated any time the requests that are made in the
tests are modified. I think that this is still better than not having
any, because this prevents silent breakage and highlights the changes
that should be manually tested against the real API.