gandiv5: Add "Bearer" prefix to the auth header
The "Bearer" prefix is required per
https://api.gandi.net/docs/authentication/ to declare the content of
the Authorization header as a personal access token.

Without this change, the user has to specify

  GANDIV5_PERSONAL_ACCESS_TOKEN='Bearer 12341234'

instead of

  GANDIV5_PERSONAL_ACCESS_TOKEN=12341234

which disagrees with the documentation and the existing test data.

This change is not backwards-compatible. This should be OK because the
commit 113648a36817a5d9667b8dd1f1c6c67eeb914916 which added support
for personal access tokens was not in any release yet.