gcloud: support GCE_ZONE_ID to bypass zone list
The GCloud IAM permission system permits a zone to grant access to an actor,
without the project granting any access.  This can be used with Service
Accounts to let an SA edit DNS in one particular zone, and nothing more.

Remove the need for the caller to have project-level role access granting the
`dns.managedZones.list` permission, in exchange for the caller telling us the
explicit zone ID to use, via the `GCE_ZONE_ID` environment variable.