feat: Change renew rc when cert is still valid
Sometimes the lego Client may fail to contact the CA due to some
temporary networking issue on the host. The certificate may otherwise
still be valid, and a user may want to ignore this issue.

This change modifies the renew service so that if the lego client
is nil, but the certificate is not expired, the exit code of the
program will be 2 instead of 1.

In order to implement this, I had to change the setup function to remove
a usage of log.Fatalf which in turn required me to introduce explicit
os.Exit calls in some commands. Although this was the lowest patch delta
solution, I also considered refactoring the setup entirely to remove
log.Fatalf, and return setup errors instead.